+++ src/PhpSpreadsheet/Worksheet/CellIterator.php
@@ -17,6 +17,10 @@ abstract class CellIterator implements NativeIterator
 
     public const TREAT_EMPTY_STRING_AS_EMPTY_CELL = 2;
 
+    public const IF_NOT_EXISTS_RETURN_NULL = false;
+
+    public const IF_NOT_EXISTS_CREATE_NEW = true;
+
     /**
      * Worksheet to iterate.
      *
@@ -39,12 +43,30 @@ abstract class CellIterator implements NativeIterator
     protected $onlyExistingCells = false;
 
     /**
+     * If iterating all cells, and a cell doesn't exist, identifies whether a new cell should be created,
+     *    or if the iterator should return a null value.
+     *
+     * @var bool
+     */
+    protected $ifNotExists = self::IF_NOT_EXISTS_CREATE_NEW;
+
+    /**
      * Destructor.
      */
     public function __destruct()
     {
         // @phpstan-ignore-next-line
         $this->worksheet = $this->cellCollection = null;
+    }
+
+    public function getIfNotExists(): bool
+    {
+        return $this->ifNotExists;
+    }
+
+    public function setIfNotExists(bool $ifNotExists = self::IF_NOT_EXISTS_CREATE_NEW): void
+    {
+        $this->ifNotExists = $ifNotExists;
     }
 
     /**